pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# Framework to make backend apps
# https://pypi.org/project/fastapi/
fastapi = "^0.115.12"
# Manager for db migrations based on sqlalchemy
# https://github.com/sqlalchemy/alembic
alembic = "1.13.2"
# Async orm for python
# https://www.sqlalchemy.org/
sqlalchemy = "^2.0.40"
# Async engine for postgres
# https://github.com/MagicStack/asyncpg
asyncpg = "^0.30.0"
# Manage data for API
# https://docs.pydantic.dev/latest/
pydantic = "^2.11.3"
# ASGI for backend
# https://github.com/encode/uvicorn
uvicorn = "0.34.1"
# Extension for password encryption
# https://passlib.readthedocs.io/en/stable/
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
# Tool to manage environment variables
# https://pypi.org/project/python-decouple/
python-decouple = "^3.8"
# Tool to generate jwt tokens
# https://pyjwt.readthedocs.io/en/stable/
pyjwt = "^2.10.1"
# Easier to manage datetime objects
# https://arrow.readthedocs.io/en/latest/
arrow = "^1.3.0"
# The best linter written on rust
# https://docs.astral.sh/ruff/
ruff = "^0.11.7"
# To handle git hooks in python-level
# https://github.com/pre-commit/pre-commit
pre-commit = "^4.3.0"
# Make routine command shorter
# https://www.pyinvoke.org/
invoke = "^2.2.0"
# Great invoke commands
# https://github.com/saritasa-nest/saritasa-invocations
saritasa-invocations = "^1.6.0"
# Framework for tests on python
# https://docs.pytest.org/en/stable/contents.html
pytest = "^8.4.1"
# Provide asynchronous mode for tests
# https://pytest-asyncio.readthedocs.io/en/stable/
pytest-asyncio = "^1.1.0"
# Provide pretty print of output for tests
# https://pypi.org/project/pytest-sugar/
pytest-sugar = "^1.1.1"
# Generate instances
# https://factoryboy.readthedocs.io/en/stable/
factory-boy = "^3.3.3"
# Http client for testing
# https://www.python-httpx.org/
httpx = "^0.28.1"
# Provide lazy loading of fixtures
# https://pypi.org/project/pytest-lazy-fixtures/
pytest-lazy-fixtures = "^1.3.4"
# Settings management using Pydantic.
# https://docs.pydantic.dev/latest/concepts/pydantic_settings/
pydantic-settings = "^2.10.1"
# Redis python client
# https://github.com/redis/redis-py
redis = "^6.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = [
    "I",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"config/__init__.py" = ["F403"]

[tool.ruff.lint.isort]
force-wrap-aliases = true
split-on-trailing-comma = true
section-order = [
    "future",
    "standard-library",
    "fastapi",
    "third-party",
    "first-party",
    "settings",
    "source",
    "local-folder",
]
[tool.ruff.lint.isort.sections]
fastapi = ["fastapi"]
settings = ["config"]
source = ["src"]

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "src"
]
addopts = [
    "--asyncio-mode=auto",
    "--ff",
    "--verbose",
    "--cache-clear"
]
asyncio_default_test_loop_scope = "session"
